+++ TagsCloudContainer/Visualisation/PngTagsCloudRenderer.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using TagsCloudContainer.TagsClouds;
+using TagsCloudContainer.Visualisation.Coloring;
 
 namespace TagsCloudContainer.Visualisation
 {
